+++ docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+### 2024-2-2 version="0.3.8"
+1. Discord、Telegram支持设置代理登录 https://stackoverflow.com/a/71668186
+
 ### 2024-2-1 version="0.3.5"
 1. 适配Telegram对发送图片行为的改动
 2. 通过Telgram平台生成的图片会保存在data/images